# Replace full-width prompt header with a variable-width-with-max prompt header