--- a/main/util/src/mill/util/PromptLoggerUtil.scala
+++ b/main/util/src/mill/util/PromptLoggerUtil.scala
@@ -150,7 +150,7 @@ private object PromptLoggerUtil {
     // For non-interactive jobs, the prompt won't be at the bottom of the terminal but
     // will instead be in the middle of a big log file with logs above and below, so we
     // need some kind of footer to tell the reader when the prompt ends and logs begin
-    val footer = Option.when(!interactive)("=" * maxWidth).toList
+    val footer = Option.when(!interactive)("=" * header.length).toList
 
     header :: body ::: footer
   }
@@ -187,28 +187,16 @@ private object PromptLoggerUtil {
     val headerPrefixStr = if (!interactive || ending) headerPrefix else s"  $headerPrefix"
     val headerSuffixStr = headerSuffix0
     val titleText = s" $titleText0 "
-    // -12 just to ensure we always have some ==== divider on each side of the title
+
+    val dividerLength = 30
     val maxTitleLength =
-      maxWidth - math.max(headerPrefixStr.length, headerSuffixStr.length) * 2 - 12
+      maxWidth - headerPrefixStr.length - headerSuffixStr.length - dividerLength * 2
     val shortenedTitle = splitShorten(titleText, maxTitleLength)
 
-    // +2 to offset the title a bit to the right so it looks centered, as the `headerPrefixStr`
-    // is usually longer than `headerSuffixStr`. We use a fixed offset rather than dynamically
-    // offsetting by `headerPrefixStr.length` to prevent the title from shifting left and right
-    // as the `headerPrefixStr` changes, even at the expense of it not being perfectly centered.
-    val leftDivider = "=" * ((maxWidth / 2) - (titleText.length / 2) - headerPrefixStr.length + 2)
-    val rightDivider =
-      "=" * (
-        maxWidth - headerPrefixStr.length - leftDivider.length -
-          shortenedTitle.length - headerSuffixStr.length
-      )
-    val headerString =
-      headerPrefixStr + leftDivider + shortenedTitle + rightDivider + headerSuffixStr
-    assert(
-      headerString.length == maxWidth,
-      s"${pprint.apply(headerString)} is length ${headerString.length}, requires $maxWidth"
-    )
-    headerString
+    val divider = "=" * dividerLength
+
+    val headerString = headerPrefixStr + divider + shortenedTitle + divider + headerSuffixStr
+    splitShorten(headerString, maxWidth)
   }
 
   def splitShorten(s: String, maxLength: Int): String = {