The previous full-width prompt header looks really weird on wide screens, even a full-width laptop screen is enough to make it awkward. This PR replaces it with a variable-width prompt header with some heuristics for shortening the `=` bars and the title text as necessary.

Covered by existing tests, added some more thorough tests in `PromptLoggerUtilTests.scala` to exercise the edge cases in the new shortening heuristics