@lihaoyi 2024-11-23
CC @lefou I think we discussed something like this before