@lefou 2024-11-23
I don't need the dashes at all. Also I accidentally thought some time ago that you changed to progress logger to have the seconds before the task name and I really liked it. Then I realized, I was looking at some evaluator command running anonymous tasks with with extra ticker output. Nevertheless, If we move the timing between the progress and the task name, we could end up with a condensed consistent output:

Before:

```
[15624/16847] ================================== __.compile ========================================== 14s
[15133] main.client.compile 3s
[15397] scalanativelib.worker-api.compile 1s
[15404] scalajslib.worker-api.compile 1s
[15612] kotlinlib.worker.test.resolvedIvyDeps
[15613] javascriptlib.test.resolvedIvyDeps
[15614] main.api.test.resolvedIvyDeps
[15615] contrib.versionfile.test.resolvedIvyDeps
[15616] contrib.sonatypecentral.test.resolvedIvyDeps
[15617] contrib.scoverage.worker2.test.resolvedIvyDeps
[15618] contrib.scoverage.api.test.resolvedIvyDeps
[15619] contrib.playlib.worker[3.0].test.resolvedIvyDeps
[15620] contrib.playlib.worker[2.9].test.resolvedIvyDeps
[15621] contrib.playlib.worker[2.8].test.resolvedIvyDeps
[15608] main.eval.test.resolvedIvyDeps
[15611] main.codesig.test.resolvedIvyDeps
[15173] contrib.testng.compile 2s
```

After (mocked):

```
[15624/16847]  14s  ==> __.compile
[15133]  3s  main.client.compile
[15397]  1s  scalanativelib.worker-api.compile
[15404]  1s  scalajslib.worker-api.compile
[15612]      kotlinlib.worker.test.resolvedIvyDeps
[15613]      javascriptlib.test.resolvedIvyDeps
[15614]      main.api.test.resolvedIvyDeps
[15615]      contrib.versionfile.test.resolvedIvyDeps
[15616]      contrib.sonatypecentral.test.resolvedIvyDeps
[15617]      contrib.scoverage.worker2.test.resolvedIvyDeps
[15618]      contrib.scoverage.api.test.resolvedIvyDeps
[15619]      contrib.playlib.worker[3.0].test.resolvedIvyDeps
[15620]      contrib.playlib.worker[2.9].test.resolvedIvyDeps
[15621]      contrib.playlib.worker[2.8].test.resolvedIvyDeps
[15608]      main.eval.test.resolvedIvyDeps
[15611]      main.codesig.test.resolvedIvyDeps
[15173]  2s  contrib.testng.compile
```