@lihaoyi 2024-11-24
For me I need some kind of dashes, otherwise i find the lack of separation makes it visually hard to distinguish log lines from prompt lines.

The timing on the left thing is possible, but one issue is the timing string is variable width, from zero to four digits wide (to give an arbitrary upper limit). So either we would need to reserve a pretty wide gutter, or we would have un-aligned task selectors, or we would have task selectors moving left and right as the longest lived tasks complete. Might be possible to find a compromise but may need more thought