@lefou 2024-12-06
> but one issue is the timing string is variable width, from zero to four digits wide 

We could use a `mm:ss` format for the timing.